Dev > Main

+++ src/main/kotlin/ar/edu/unsam/phm/dto/CreateReservationDTO.kt
@@ -1,11 +1,10 @@
 package ar.edu.unsam.phm.dto
 
-import ar.edu.unsam.phm.domain.Reservation
 import java.time.LocalDate
 
 data class CreateReservationDTO(
-    val bookId: Int,
-    val sessionId: Int,
+    val bookId: Long,
+    val sessionId: Long,
     val pickUpDate: LocalDate,
     val dropOffDate: LocalDate
-){}
+) {}
